Make minor changes in models and work on course manager.

+++ MyLearn/MyLearn/Models/ReturnCode.cs
@@ -2,6 +2,6 @@
 {
     public class ReturnCode
     {
-        public float ReturnStatus { get; set; }
+        public int ReturnStatus { get; set; }
     }
 }
